sandb/requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.1.5
django-crispy-forms==1.14.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
Werkzeug==2.2.2
